moving to multidimensional arrays

+++ PostMessage.php
@@ -12,18 +12,37 @@ ini_set('display_errors', 'On');
             echo "<p>All fields must be filled.</p>";
             echo "<p><a href=\"DiscussionForum.php\">Go Back.</a></p>";
         } else {
+
             $Topic = $_POST['topic'];
             $Name = $_POST['name'];
             $Message = $_POST['message'];
-    
+
             $PostMessage = addslashes("$Topic~$Name~$Message\n");
-            $MessageStore = fopen("messages.txt","a+");
-            fwrite($MessageStore, "$PostMessage");
-            fclose($MessageStore);
-    
-            echo "<p><strong>Topic</strong>: $Topic<br/></p>";
-            echo "<p><strong>Name</strong>: $Name<br/></p>";
-            echo "<p><strong>Message</strong>:$Message</p>";
+
+            $TopicExists = FALSE;
+            if (file_exists("messages.txt") && filesize("messages.txt") > 0) {
+                $MessageArray = file("messages.txt");
+                for ($i = 0; $i < count($MessageArray); ++$i) {
+                    $CurMessage = explode("~", $MessageArray[$i]);
+                    if (in_array(addslashes($Topic), $CurMessage)) {
+                        $TopicExists = TRUE;
+                        break;
+                    }
+                }
+            }
+
+            if ($TopicExists) {
+                echo "<p>The topic already exists.</p>";
+            } else {
+
+                $MessageStore = fopen("messages.txt","a+");
+                fwrite($MessageStore, "$PostMessage");
+                fclose($MessageStore);
+        
+                echo "<p><strong>Topic</strong>: $Topic<br/></p>";
+                echo "<p><strong>Name</strong>: $Name<br/></p>";
+                echo "<p><strong>Message</strong>:$Message</p>";
+            }
         }
     ?>
     <hr>
